rename: 파일명 명확하게 변경 및 import 수정
- github.ts 이름 중복 없앰
- IssueList.tsx 파일의 import 수정
